python-client/requirements.txt
spotipy==2.25.2
google-api-python-client==2.164.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-socketio[client]==5.12.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-mock==3.14.0
